Add an environment variable to enable HTTP2
HTTP2 had many issues in Prometheus. Before enabling it again, because
some of the underlying issues have been fixed, I would like to introduce
an environment variable to enable it.

That would enable us to contact some "power users" to test this in the
conditions where they had issues to see if this can be re-enabled in
Prometheus generally.

Signed-off-by: Julien Pivotto <roidelapluie@inuits.eu>